Backend-Legacy-master_2.0/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
blinker==1.5
click==8.1.3
colorama==0.4.5
decouple==0.0.7
Flask==2.1.3
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.14.0
Mako==1.2.2
MarkupSafe==2.1.1
psycopg-binary==3.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.1.2
WTForms==3.0.1
